+++ R/utils.R
@@ -5,6 +5,7 @@
 #'
 #' @noRd
 format_icd_codes <- function(codes) {
+  codes <- strip_surrounding_quotes(codes)
   return(trimws(as.character(codes)))
 }